add explicit type case for rp

+++ examples/pubsub/pubsub_interrupt_publisher.c
@@ -276,10 +276,10 @@ UA_PubSubManager_addRepeatedCallback(UA_Server *server,
         return UA_STATUSCODE_BADINTERNALERROR;
     }
 
-    timerspec.it_interval.tv_sec        = 0;
-    timerspec.it_interval.tv_nsec       = pubIntervalNs;
-    timerspec.it_value.tv_sec           = 0;
-    timerspec.it_value.tv_nsec          = pubIntervalNs;
+    timerspec.it_interval.tv_sec        = (long int) 0;
+    timerspec.it_interval.tv_nsec       = (long int) pubIntervalNs;
+    timerspec.it_value.tv_sec           = (long int) 0;
+    timerspec.it_value.tv_nsec          = (long int) pubIntervalNs;
     resultTimerCreate                   = timer_settime(pubEventTimer, 0,
                                                         &timerspec, NULL);
     if (resultTimerCreate != 0)
@@ -303,10 +303,10 @@ UA_PubSubManager_changeRepeatedCallbackInterval(UA_Server *server,
     struct itimerspec timerspec;
     int               resultTimerCreate = 0;
     pubIntervalNs                       = (UA_Int64) (interval_ms * MILLI_SECONDS);
-    timerspec.it_interval.tv_sec        = 0;
-    timerspec.it_interval.tv_nsec       = pubIntervalNs;
-    timerspec.it_value.tv_sec           = 0;
-    timerspec.it_value.tv_nsec          = pubIntervalNs;
+    timerspec.it_interval.tv_sec        = (long int) 0;
+    timerspec.it_interval.tv_nsec       = (long int) pubIntervalNs;
+    timerspec.it_value.tv_sec           = (long int) 0;
+    timerspec.it_value.tv_nsec          = (long int) pubIntervalNs;
     resultTimerCreate                   = timer_settime(pubEventTimer, 0,
                                                         &timerspec, NULL);
     if (resultTimerCreate != 0)
